Code For Inspiration
The code otl.py is one of the things I will take inspiration from as I build Wagon vs Wild.
The exact name of the code is "OriginCodeAcademy" and is linked within my README file. I have run and tested the code to make sure it works.

Furthermore, I have updated the project roadmap to reflect this progress.