Finished Zombie Combat
I have just finished my combat for zombies, I still need to add the lazer rifle into the mix, however this is looking great.

For testing purposes, I have assigned the rifle and spear to be added to the player's inventory in option C.

For the rifle and spear combat I have decided to go with random encounters, these will be expanded as time goes on.

You might notice the events list has three different items, Bitten, Scratched and Notice. Bitten is instant death for the player. Scratched is going to utilize the future medical function within the game, where a med kit can be purchased to heal. Notice is supposed to be the player character reacting to the zombies before they can react to you.
However, I have erroneously added the "First Aid Function" which will be removed in the future.

I am using the further testing file to be absolutely sure the code is working before implementing it within the main code.