Added River Crossing Event
This event hinges on the player having a spare wheel within the inventory, having this item will allow the player to continue.

Notice random_event, this is sort of a roll based encounter, where the player will need to get lucky to pass, you can technically pass this stage without a spare wheel, however if your roll is unlucky then the game ends.