I have just completed the Dinosaur Event
Instead of using the old combat model used for the zombies I decided to mix things up a bit.

Keep in mind I am still using Rifle and Spear Combat, However, I've specialized a combat system for the dino event where a random number is rolled which counts as damage to the HP of both the player and the dino's.

If time permits, more will be added to the dino encounter.