I have finished the Mountain Encounter
In the mountains I have set up two different random paths, you either run into a trader or have a dangerous crossing with two branches, either move over a bridge or narrow path.
The players decision on the bridge will either result in death OR you move across and through the mountains without event.

The other branch, the narrow path results in a land slide, if you have a pick axe then you move through unscathed, without a pickaxe, means back tracking.
This results in bandits, you can either bribe the bandits OR fight them.

Keep in mind that fighting the bandits may result in the loss of a weapon. If you run into zombies again this could be a problem because the player would be down to a single weapon.

Also, I have not balanced any of the different paths at the beginning, meaning that it does not matter if you choose A, B or C at this time.

The progress I have made thus far is encouraging.

The further testing file is where I plan out the encounters and fights. You will notice that I frequently delete what is within after I am done testing it.